CLN merge simulated and quadratic
